Relative lib files and code cleaning.
Library files are now in the libraries project folder.

I also moved the dll folder into the debug, so it shouldn't require you
to move it across all the time.

+++ Project/GP2Coursework/headers/GameObject.h
@@ -49,9 +49,7 @@ public:
 	Light* getLight();
 
 	int getChildCount();
-	GameObject * getChild(int index);
-
-    
+	GameObject * getChild(size_t index);
 
 protected:
 private:
